requirements.txt
torch>=2.1.0
transformers>=4.44.0
peft>=0.12.0
trl>=0.9.0
datasets>=2.20.0
bitsandbytes>=0.43.0
accelerate>=0.33.0
rich>=13.0.0
psutil>=5.9.0
pyyaml>=6.0
